Rewritten class chooser and button handling

+++ Inc/ffb_defs.h
@@ -85,7 +85,7 @@
 #ifdef __cplusplus
 struct  __attribute__((__packed__)) reportHID_t {
 		uint8_t id = 1;
-		int16_t buttons = 0;
+		uint32_t buttons = 0;
 		int16_t X = 0;
 		int16_t Y = 0;
 		int16_t Z = 0;
